requirements.txt
absl-py
blobfile
cachetools
certifi
charset-normalizer
clean-fid
contextlib2
filelock
google-auth
google-auth-oauthlib
grpcio
idna
lmdb
lxml
Markdown
MarkupSafe
ml-collections
mpi4py
numpy
oauthlib
opencv-python
Pillow
protobuf
pyasn1
pyasn1-modules
pycryptodomex
PyYAML
requests
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
torch
torchvision
tqdm
typing_extensions
urllib3
Werkzeug
